% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{correlation Function}
\usage{
correlation(Year)
}
\arguments{
\item{Year}{WHich specific year do you want to look into?}
}
\description{
This function allows you to calculate correlation between numerical data in any year.
}
\examples{
highest_score()
}
\keyword{NBA}
