% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_score.R
\name{highest_score}
\alias{highest_score}
\title{Highest Score Function}
\usage{
highest_score(Year)
}
\arguments{
\item{Year}{WHich specific year do you want to look into?}
}
\description{
This function allows you to know the highest score of one NBA player in any year.
}
\examples{
highest_score()
}
\keyword{NBA}
