% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(Year)
}
\arguments{
\item{Year}{Which specific year do you want to look into?}
}
\description{
This function allows you to know the oldest NBA player in any year.
}
\examples{
highest_score()
}
\keyword{NBA}
